requirements.txt
alembic==1.11.1 ; python_version >= "3.10" and python_version < "4.0"
anyio==3.7.0 ; python_version >= "3.10" and python_version < "4.0"
async-timeout==4.0.2 ; python_version >= "3.10" and python_full_version <= "3.11.2"
blinker==1.6.2 ; python_version >= "3.10" and python_version < "4.0"
certifi==2023.5.7 ; python_version >= "3.10" and python_version < "4.0"
cfgv==3.3.1 ; python_version >= "3.10" and python_version < "4.0"
click==8.1.3 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0" and sys_platform == "win32" or python_version >= "3.10" and python_version < "4.0" and platform_system == "Windows"
decorator==5.1.1 ; python_version >= "3.10" and python_version < "4.0"
distlib==0.3.6 ; python_version >= "3.10" and python_version < "4.0"
dnspython==2.3.0 ; python_version >= "3.10" and python_version < "4.0"
email-validator==2.0.0.post2 ; python_version >= "3.10" and python_version < "4.0"
exceptiongroup==1.1.1 ; python_version >= "3.10" and python_version < "3.11"
faker==18.9.0 ; python_version >= "3.10" and python_version < "4.0"
fastapi-pagination==0.12.4 ; python_version >= "3.10" and python_version < "4.0"
fastapi==0.95.2 ; python_version >= "3.10" and python_version < "4.0"
fastapi[all]==0.95.2 ; python_version >= "3.10" and python_version < "4.0"
filelock==3.12.0 ; python_version >= "3.10" and python_version < "4.0"
greenlet==2.0.2 ; python_version >= "3.10" and platform_machine == "aarch64" and python_version < "4.0" or python_version >= "3.10" and platform_machine == "ppc64le" and python_version < "4.0" or python_version >= "3.10" and platform_machine == "x86_64" and python_version < "4.0" or python_version >= "3.10" and platform_machine == "amd64" and python_version < "4.0" or python_version >= "3.10" and platform_machine == "AMD64" and python_version < "4.0" or python_version >= "3.10" and platform_machine == "win32" and python_version < "4.0" or python_version >= "3.10" and platform_machine == "WIN32" and python_version < "4.0"
gunicorn==20.1.0 ; python_version >= "3.10" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.10" and python_version < "4.0"
httpcore==0.17.2 ; python_version >= "3.10" and python_version < "4.0"
httptools==0.5.0 ; python_version >= "3.10" and python_version < "4.0"
httpx==0.24.1 ; python_version >= "3.10" and python_version < "4.0"
identify==2.5.24 ; python_version >= "3.10" and python_version < "4.0"
idna==3.4 ; python_version >= "3.10" and python_version < "4.0"
itsdangerous==2.1.2 ; python_version >= "3.10" and python_version < "4.0"
jinja2==3.1.2 ; python_version >= "3.10" and python_version < "4.0"
kafka-python==2.0.2 ; python_version >= "3.10" and python_version < "4.0"
loguru==0.7.0 ; python_version >= "3.10" and python_version < "4.0"
mako==1.2.4 ; python_version >= "3.10" and python_version < "4.0"
markupsafe==2.1.2 ; python_version >= "3.10" and python_version < "4.0"
nodeenv==1.8.0 ; python_version >= "3.10" and python_version < "4.0"
orjson==3.8.14 ; python_version >= "3.10" and python_version < "4.0"
pinject==0.14.1 ; python_version >= "3.10" and python_version < "4.0"
platformdirs==3.5.1 ; python_version >= "3.10" and python_version < "4.0"
pre-commit==3.3.2 ; python_version >= "3.10" and python_version < "4.0"
psycopg2-binary==2.9.6 ; python_version >= "3.10" and python_version < "4.0"
pydantic==1.10.8 ; python_version >= "3.10" and python_version < "4.0"
pyjwt==2.7.0 ; python_version >= "3.10" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.10" and python_version < "4.0"
python-dotenv==1.0.0 ; python_version >= "3.10" and python_version < "4.0"
python-multipart==0.0.6 ; python_version >= "3.10" and python_version < "4.0"
pyyaml==6.0 ; python_version >= "3.10" and python_version < "4.0"
redis==4.5.5 ; python_version >= "3.10" and python_version < "4.0"
setuptools==67.8.0 ; python_version >= "3.10" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
sniffio==1.3.0 ; python_version >= "3.10" and python_version < "4.0"
sqlalchemy==2.0.15 ; python_version >= "3.10" and python_version < "4.0"
starlette==0.27.0 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.6.2 ; python_version >= "3.10" and python_version < "4.0"
ujson==5.7.0 ; python_version >= "3.10" and python_version < "4.0"
uvicorn[standard]==0.22.0 ; python_version >= "3.10" and python_version < "4.0"
uvloop==0.17.0 ; sys_platform != "win32" and sys_platform != "cygwin" and platform_python_implementation != "PyPy" and python_version >= "3.10" and python_version < "4.0"
virtualenv==20.23.0 ; python_version >= "3.10" and python_version < "4.0"
watchfiles==0.19.0 ; python_version >= "3.10" and python_version < "4.0"
websockets==11.0.3 ; python_version >= "3.10" and python_version < "4.0"
win32-setctime==1.1.0 ; python_version >= "3.10" and python_version < "4.0" and sys_platform == "win32"
